refactor(curator): extract LLM review fork into agent/curator_review_fork.py
The consolidation-guard fix in this branch pushed agent/curator.py past
this repo's 2,000-line file-size gate (2101 lines). Extracts the one part
of the curator that actually talks to a model — CURATOR_REVIEW_PROMPT /
CURATOR_DRY_RUN_BANNER, provider/model resolution
(_resolve_review_runtime / _resolve_review_model), and the forked-AIAgent
construction and execution (_run_llm_review, including the toolset
restriction and dispatch-time whitelist wiring) — into its own bounded
module. curator.py keeps scheduling, state, deterministic lifecycle
transitions, and report writing/rendering.

curator.py re-exports the moved names so every existing `curator.X` /
`from agent.curator import X` reference, including test monkeypatches of
_run_llm_review, keeps working unchanged. The new module's logger is
explicitly bound to "agent.curator" (not __name__) so log filtering and
caplog-based tests keyed on that logger name see no change.

agent/curator.py: 2101 -> 1603 lines. New agent/curator_review_fork.py:
538 lines. Both under the 2,000-line gate.
